pyproject.toml
[project]
name = "url-shortener"
version = "0.1.0"
description = "URL Shortener service built with FastAPI"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic-settings>=2.1.0",
    "pytest>=7.4.4",
    "httpx>=0.26.0",
]
requires-python = ">=3.12"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]